+++ plugins/maxim/changelog.md
@@ -1,4 +1,5 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- chore: version update core to 1.2.14 and framework to 1.1.16
+- chore: version update core to 1.2.15 and framework to 1.1.17
+- feat: added support for streaming responses